{"schema":"decodex/commit/1","summary":"Split account panel and review lifecycle modules by responsibility","authority":"manual"}
